extract test first leaf
