+++ package.json
@@ -11,7 +11,7 @@
     "test:node": "hundreds mocha test/test-*.js",
     "test:browser": "polendina --cleanup dist/cjs/node-test/test-*.js --webpack-config test/polendina-webpack.config.cjs",
     "test": "npm run lint && npm run test:node && npm run test:cjs && npm run test:browser",
-    "coverage": "c8 --reporter=html mocha test/test-*.js -- --grep specific && npm_config_yes=true npx st -d coverage -p 8080"
+    "coverage": "c8 --reporter=html mocha test/test-*.js -- --grep inline && npm_config_yes=true npx st -d coverage -p 8080"
   },
   "keywords": [],
   "author": "Mikeal Rogers <mikeal.rogers@gmail.com> (https://www.mikealrogers.com/)",
